instablog-jarang/requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
Django==1.9.2
django-bootstrap3==7.0.0
djangorestframework==3.3.3
kombu==3.0.34
Pillow==3.1.1
pytz==2015.7
redis==2.10.5
